+++ CHANGELOG.md
@@ -8,7 +8,8 @@ Go to the `v1` branch to see the changelog of Lume 1.
 
 ## [3.0.10] - Unreleased
 ### Added
-- New flag `--proxied`, combined with `--serve` to use a location with a different hostname and port than the server's hostname and port.
+- New flag `--proxied` combined with `--serve` to use a location with a different hostname and port to the server's hostname and port.
+- Ignore `deno.json` and `deno.jsonc` files by the watcher.
 
 ### Fixed
 - `icons` plugin: clear the plugin list on update.