+++ CHANGELOG.md
@@ -6,10 +6,10 @@ and this project try to adheres to [Semantic Versioning](https://semver.org/).
 Go to the `v2` branch to see the changelog of Lume 2.
 Go to the `v1` branch to see the changelog of Lume 1.
 
-## [3.0.10] - Unreleased
+## [3.0.10] - 2025-09-24
 ### Added
 - New env variable `LUME_PROXIED` to use a location with a different hostname and port to the server's hostname and port.
-- New file `serve.ts` to run LumeCMS in a VPS.
+- New file `serve.ts` to run LumeCMS in a VPS (experimental, will be stable in 3.1).
 - Ignore `deno.json` and `deno.jsonc` files by the watcher.
 
 ### Fixed
@@ -248,7 +248,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 [#779]: https://github.com/lumeland/lume/issues/779
 [#782]: https://github.com/lumeland/lume/issues/782
 
-[3.0.10]: https://github.com/lumeland/lume/compare/v3.0.9...HEAD
+[3.0.10]: https://github.com/lumeland/lume/compare/v3.0.9...v3.0.10
 [3.0.9]: https://github.com/lumeland/lume/compare/v3.0.8...v3.0.9
 [3.0.8]: https://github.com/lumeland/lume/compare/v3.0.7...v3.0.8
 [3.0.7]: https://github.com/lumeland/lume/compare/v3.0.6...v3.0.7